Make dicom server status screen prettier
Summary:
http://dicom-server.eng.aws.grail.com:8080/datasets

Fix a problem in the dicom parser where having a PixelData in the middle of a
sequence or an item crashes the process, when DropPixelData flag is set.
The problem is that we don't fully consume the sequence, then try to parse the
rest of the file which becomes garbage.

Run two HTTP servers in the server, one public facing (and serves an empty page)
and the one that's for real. The public page can be used for connectivity
testing by partners.

Reviewers: jbredno

Reviewed By: jbredno

Differential Revision: https://phabricator.grailbio.com/D9372

fbshipit-source-id: e2ff811