Refactor API, add Parser Interface (#13)
This change refactors the core `go-dicom` API to use an interface called `Parser`. This allows for more elegant dependency injection of items (like channels, previous data elements) with a clean end user api (so that the user doesn't have to worry about passing in those references themselves). Using an interface like this will also make typical testing easier on the end user, and is also more canonical golang imho. This closes #11 and addresses #3. This also puts some plumbing in place for #5.