Refactor element int api (#36)
This change refactors the API used to extract integer values from a DICOM Element. A top level API GetInt and GetInts retrieve the stored integer-like value (which may be uint16, int8, etc) consistently as either an int64 or []int64.

If the caller wants to access the "true" value of the integer(s), the caller can do so by directly accessing Element.Value and type casting as needed based on the Element's VR.

All legacy methods to retrieve specific types of ints have been removed.

There is some small-in-practice efficiency trade off to be typecasting these integers, but this API is clearer and easier to use and cleans up the code a fair bit. Power users can still easily work with the Values directly if efficiency is of the utmost importance.

This chance also factors out DataSet into its own file in the element package.