[Rewritten Prototype] Implement dicomio.Reader, add simple test program, fix some bugs (#70)
This change implements the initial dicomio.Reader, which unblocks the first end-to-end test of the rewritten prototype parser. A simple test program was added to test the library, which appears to successfully parse several simple test DICOMs!

Several bugs were also caught and fixed.

More robust e2e testing against a suite of public DICOMs should be added, in addition to more comprehensive unit testing as this rewritten library begins to take shape.

Since this is on the rewrite branch, going to submit this to continue moving forward and building out a prototype rewrite.