Support reading non-standard DICOMs without magic word (#127)
This change adds some support to read non-standard DICOMs that do not include the magic word at offset 128 (as described in #59).

This change attempts to peek ahead at the magic word to see if it is there. If not, it effectively rewinds back to byte zero, skips metadata parsing, and goes back to typical parsing (with a transfer syntax default of LittleEndian Implicit).

While some of the non-standard dicoms in #59 appear to be broken in other unrelated ways, most of them work with this fix.