write: Item VL should not default to UndefinedLength (#182)
I do not think that we should be writing out item VLs as undefined length by default. However, if the caller calls writeVRVL with an item and provides an undefined length VL, then that will be respected (which is done for non-encapsulated image sequences).

__Unlike other dicom sequences__, the encapsulated image sequence items cannot be written with undefined length. 

> The encapsulated pixel stream of encoded pixel data is segmented into one or more Fragments, __each of which conveys its own explicit length__. The sequence of Fragments of the encapsulated pixel stream is terminated by a delimiter, thus allowing the support of encoding processes where the resulting length of the entire pixel stream is not known until it is entirely encoded.

http://dicom.nema.org/dicom/2013/output/chtml/part05/sect_8.2.html

The effective change here is that all other sequence items will continue to be written out with undefined length VLs (and Item delimitation items), except for encapsulated image sequences.