Introduce FlatStatefulIterator, and ExhaustElementChannel Helper (#184)
This addresses the discussion in #183 by introducing a FlatStatefulIterator (that will be GC'd normally, even if it isn't exhausted). This also introduces ExhaustElementChannel which is an easy way for folks using the channel API to ensure it is always exhausted.

At some point either we should deprecate the channel-based API, or we should rename it so that it's not seen as the default for iteration (e.g. name it FlatChannelIterator or something). For most usecases, the stateful iterator should be sufficient.

Future improvements can we made to the stateful iterator, for example not recomputing the flat element representation if the dataset hasn't changed from a previous call.