Do not replace caller-supplied VRs with UN for unknown tags, even when verifyVROrDefault is enabled (#194)
This PR closes #193.

This PR changes the behavior of the writer to default to any existing VR value when we do not have any information on the tag, effectively trusting that the caller knows more than we do about this element. I've added a couple tests to verify this is working as intended, and I've also refactored the logic a bit in verifyVROrDefault and added some code comments explaining the intent of certain logic paths.