Allow option to tolerate PixelData VL and expected length mismatches. (#245)
This change introduces a ParseOption that allows PixelData element VL and expected lengths to mismatch without returning an error. If the given ParseOption is set, the PixelData with the aforementioned mismatch will have its raw bytes read, dropped into an encapsulated Frame, and have that Frame wrapped in a PixelDataInfo with a ParseErr property set to an error describing this scenario. 

(Commit description/summary set by suyashkumar@).

Co-authored-by: quang18 <quang@segmed.ai>