Deflate transfer syntax support (for parse) (#321)
This adds support for Deflated Little Endian Explicit transfer syntax on parsing. Support on writing should come in another change.

This change does this by adding a one way SetDeflate mechanism to the dicomio.Reader. The expectation for now is that this internal API is called once and at the top level of dicom parsing (e.g this doesn't support parsing just some sub-payloads with deflation, though we can make some changes to support that in the future if it's a need).