Add WriteOption to override missing transfer syntax in dataset (#332)
This fixes #328. This change will also help set the stage for roundtrip tests for transfer syntax uid inference in the case of missing transfer syntax UIDs (part of #327).

For the future:

should we give this option as something to always override whatever transfer syntax is present (or if not present, use the override as well)?