Allow unknown specific character set to default to UTF-8 encoding when parsing (#346)
This change implements an option to allow an unknown specific character set when parsing a DICOM and fallback to using UTF-8.

For context, in my application of this parser, we are seeing erroneous specific character sets such as
```
ISO IR 192
ISO_2022_IR_6
NONE
```
Current behaviour is that parsing fails immediately upon encountering an unknown character set, and the error also indicates that it will fallback to utf-8 (which is misleading since it just errors out instead of actually falling back). It seems there was a `TODO` in place for this, so this change implements the option to allow this.